+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.7.0 (2024-06-11)
+* Add the `canu.show_history_page` method.
+* Add the `canu.EventHandler.redundant` attribute.
+* Add the `canu.handle_files` method.
+* Add the `canu.Container.code_interpreter_files` attribute.
+
 ## 0.6.0 (2024-06-09)
 * Add the `canu.EventHandler.on_event` method.
 * Add the `canu.EventHandler.on_tool_call_delta` method.